Refactoring TTIR passes from single file to couple of smaller files
